Making the medulla drivers and robot state compatible with the wider error flags.

+++ software/atrias_medulla_drivers/include/atrias_medulla_drivers/HipMedulla.h
@@ -28,7 +28,7 @@ class HipMedulla : public Medulla {
 	uint8_t*  id;
 	uint8_t*  state;
 	uint8_t*  timingCounter;
-	uint8_t*  errorFlags;
+	uint16_t* errorFlags;
 	uint8_t*  limitSwitches;
 	
 	uint32_t* hipEncoder;
